+++ Icarus/core/mcp/types/input_schema.py
@@ -1,29 +1,7 @@
 
-from Icarus.core.mcp.types.mcp_property import MCPProperty
+from core.mcp.types.mcp_property import MCPProperty
 from dataclasses import dataclass
 
 @dataclass
 class InputSchema:
-    name: str
-    desc: str
-    IS_type: object
     properties: list[MCPProperty]
-    required: list[str]
-    schema = { # REMOVE THIS - ONLY A TEMPLATE
-        "inputSchema": {
-            "type": "object",
-            "properties": {
-                "location": {
-                    "type": "string",
-                    "description": "The city and state/country, e.g., 'San Francisco, CA' or 'Tokyo, Japan'"
-                },
-                "unit": {
-                    "type": "string",
-                    "enum": ["celsius", "fahrenheit"],
-                    "default": "celsius",
-                    "description": "The temperature unit to return."
-                }
-            },
-            "required": ["location"]
-        }
-    }