+++ Icarus/skills/get_time/execute.py
@@ -2,4 +2,4 @@
 import time
 
 def execute() -> str:
-    return f"The time is {time.strftime('%H %M')}"
+    return f"The time is {time.strftime('%H:%M')}"