+++ Icarus/skills/use_durapy_function/execute.py
@@ -1,3 +1,0 @@
-
-def execute() -> str:
-    return "I am not yet capable of using functions from DuraPy, but I will be in the future."